Implement 'IExClosureAttr' interface for ExClosure and ExNativeClosure
